+++ cpp_server/include/RobotEPuck.h
@@ -12,7 +12,7 @@ class RobotEPuck : public RobotBase, public Enki::EPuck
         RobotEPuck( std::string, unsigned );
         ~RobotEPuck();
         virtual void controlStep( double dt );
-        virtual void getSensors( JSON& resp );
+        virtual void getSensors( Json::Value& resp );
         virtual void setLeds( double*, int );
         virtual unsigned char* getCameraImage( unsigned int* );